Serialize the ActionReferenceForm as a uint
This removes a few casts.

+++ src/PSFilterShim/PSApi/PICA/Action Suites/ActionReferenceSuite.cs
@@ -240,7 +240,7 @@ namespace PSFilterLoad.PSApi.PICA
                 ActionReferenceItem item = container.GetReference();
                 if (item != null)
                 {
-                    *value = (uint)item.Form;
+                    *value = item.Form;
 
                     return PSError.kSPNoError;
                 }
